// road-rail-crossing-ng/src/app/model/road-rail-crossing.ts
export class RoadRailCrossing {
    name: string;
    railName: string;
    roadName: string;
    safetyLevel: string;
    vehiclePerDay: number;
    trainPerDay: number;
    mediumLengthOfVehicle: number;
    criticalDistanceLengthOfVehicle: number;
    mediumVelocityOfVehicle: number;
    mediumLengthOfTrain: number;
    criticalDistanceLengthOfTrain: number;
    mediumVelocityOfTrain: number;
    numberOfYearsOfAccidents: number;
    totalNumberOfAccidents: number;
}